fix: Rename enty file to OfflineClient
